+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -1,10 +1,9 @@
 package chess.domain.board;
 
 import java.util.Map;
-import chess.domain.Blank;
-import chess.domain.MovePath;
-import chess.domain.Piece;
-import chess.domain.PieceColor;
+import chess.domain.piece.Blank;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceColor;
 
 public class ChessBoard {
