+++ src/main/java/chess/domain/dao/GameDaoImpl.java
@@ -2,7 +2,7 @@ package chess.domain.dao;
 
 import chess.domain.dto.GameDto;
 import chess.domain.entity.Game;
-import chess.domain.util.DbConnection;
+import chess.util.DbConnection;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
