+++ src/main/java/chess/domain/dto/BoardDto.java
@@ -1,6 +1,6 @@
 package chess.domain.dto;
 
-import chess.domain.Piece;
+import chess.domain.piece.Piece;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
 import java.util.HashMap;
