+++ src/main/java/chess/domain/dto/MovableDto.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain.dto;
 
 public class MovableDto {
