+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,26 +1,26 @@
-package chess.domain;
+package chess.domain.piece;
 
 import java.util.List;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
 
-public class Knight extends Piece {
+public class Bishop extends Piece {
 
-    public Knight(PieceColor pieceColor, Position position) {
-        super(PieceType.KNIGHT, pieceColor, position);
+    public Bishop(PieceColor pieceColor, Position position) {
+        super(PieceType.BISHOP, pieceColor, position);
     }
 
     @Override
     public Direction findByDirection(Position from, Position to) {
-        final int column = to.getColumn() - from.getColumn();
+        final int colum = to.getColumn() - from.getColumn();
         final int row = to.getRow() - from.getRow();
 
-        return Direction.of(column, row);
+        return Direction.ofDiagonal(colum, row);
     }
 
     @Override
     protected List<Direction> findByMovableDirection(Piece piece, Direction direction) {
-        return Direction.KNIGHT_DIRECTION;
+        return Direction.DIAGONAL_DIRECTION;
     }
 
     @Override
