+++ src/main/java/chess/domain/piece/King.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 import java.util.List;
 import chess.domain.board.Direction;
