+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.board.MovePath;
 import java.util.List;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
