+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
