+++ src/test/java/chess/domain/game/ScoreTest.java
@@ -3,7 +3,7 @@ package chess.domain.game;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.PieceColor;
+import chess.domain.piece.PieceColor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import chess.domain.board.ChessBoard;
