+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -7,11 +7,7 @@ import static chess.domain.board.Direction.NORTH_WEST;
 import static chess.domain.board.Direction.SOUTH_EAST;
 import static chess.domain.board.Direction.SOUTH_WEST;
 
-import chess.domain.Bishop;
-import chess.domain.Blank;
-import chess.domain.MovePath;
-import chess.domain.Piece;
-import chess.domain.PieceColor;
+import chess.domain.board.MovePath;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
