+++ src/test/java/chess/domain/piece/KingTest.java
@@ -11,11 +11,7 @@ import static chess.domain.board.Direction.SOUTH_EAST;
 import static chess.domain.board.Direction.SOUTH_WEST;
 import static chess.domain.board.Direction.WEST;
 
-import chess.domain.Blank;
-import chess.domain.King;
-import chess.domain.MovePath;
-import chess.domain.Piece;
-import chess.domain.PieceColor;
+import chess.domain.board.MovePath;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
