+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -7,16 +7,11 @@ import static chess.domain.board.Direction.NORTH_WEST;
 import static chess.domain.board.Direction.SOUTH;
 import static chess.domain.board.Direction.SOUTH_EAST;
 import static chess.domain.board.Direction.SOUTH_WEST;
-import static chess.domain.PieceColor.BLACK;
-import static chess.domain.PieceColor.NONE;
-import static chess.domain.PieceColor.WHITE;
+import static chess.domain.piece.PieceColor.BLACK;
+import static chess.domain.piece.PieceColor.NONE;
+import static chess.domain.piece.PieceColor.WHITE;
 
-import chess.domain.Blank;
-import chess.domain.MovePath;
-import chess.domain.Pawn;
-import chess.domain.Piece;
-import chess.domain.PieceColor;
-import chess.domain.Rook;
+import chess.domain.board.MovePath;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
