+++ src/test/java/chess/domain/state/FinishedTest.java
@@ -2,7 +2,7 @@ package chess.domain.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static chess.domain.PieceColor.BLACK;
+import static chess.domain.piece.PieceColor.BLACK;
 
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
